# openapi_formatted.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ncloud2API",
        "version": "0.2.0"
    },
    "paths": {
...
        "/api/sales-shipments": {
            "get": {
                "tags": [
                    "shipments"
                ],
                "summary": "Api Shipments List",
                "operationId": "api_shipments_list_api_sales_shipments_get",
                "parameters": [
                    {
                        "name": "dates",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Dates"
                        }
                    },
                    {
                        "name": "datee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Datee"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "State"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "default": 20,
                            "title": "Rows"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipmentListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "shipments"
                ],
                "summary": "Api Create Shipment",
                "operationId": "api_create_shipment_api_sales_shipments_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WriteOperationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sales-shipments/{dh}": {
            "get": {
                "tags": [
                    "shipments"
                ],
                "summary": "Api Shipment Detail",
                "operationId": "api_shipment_detail_api_sales_shipments__dh__get",
                "parameters": [
                    {
                        "name": "dh",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Dh"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipmentDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "shipments"
                ],
                "summary": "Api Update Shipment",
                "operationId": "api_update_shipment_api_sales_shipments__dh__put",
                "parameters": [
                    {
                        "name": "dh",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Dh"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateShipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WriteOperationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sales-shipments/{dh}/audit": {
            "post": {
                "tags": [
                    "shipments"
                ],
                "summary": "Api Audit Shipment",
                "operationId": "api_audit_shipment_api_sales_shipments__dh__audit_post",
                "parameters": [
                    {
                        "name": "dh",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Dh"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuditActionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WriteOperationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuditAction": {
                "type": "string",
                "enum": [
                    "audit",
                    "unaudit",
                    "void"
                ],
                "title": "AuditAction"
            },
            "AuditActionRequest": {
                "properties": {
                    "action": {
                        "$ref": "#/components/schemas/AuditAction"
                    }
                },
                "type": "object",
                "required": [
                    "action"
                ],
                "title": "AuditActionRequest"
            },
...
            "CreateShipmentRequest": {
                "properties": {
                    "customer_id": {
                        "type": "string",
                        "title": "Customer Id",
                        "description": "\u5ba2\u6237\u7f16\u53f7 (ERP: khid)"
                    },
                    "shipment_date": {
                        "type": "string",
                        "title": "Shipment Date",
                        "description": "\u53d1\u8d27\u65e5\u671f YYYY-MM-DD HH:mm:ss"
                    },
                    "warehouse": {
                        "type": "string",
                        "title": "Warehouse",
                        "description": "\u53d1\u8d27\u4ed3\u5e93\u7f16\u53f7 (ERP: ck)"
                    },
                    "customer_addr": {
                        "type": "string",
                        "title": "Customer Addr",
                        "default": ""
                    },
                    "shipping_addr": {
                        "type": "string",
                        "title": "Shipping Addr",
                        "default": ""
                    },
                    "shipping_tel": {
                        "type": "string",
                        "title": "Shipping Tel",
                        "default": ""
                    },
                    "shipping_method": {
                        "type": "string",
                        "title": "Shipping Method",
                        "default": ""
                    },
                    "delivery_person": {
                        "type": "string",
                        "title": "Delivery Person",
                        "default": ""
                    },
                    "tracking_no": {
                        "type": "string",
                        "title": "Tracking No",
                        "default": ""
                    },
                    "freight": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Freight"
                    },
                    "salesperson": {
                        "type": "string",
                        "title": "Salesperson",
                        "default": ""
                    },
                    "contact_person": {
                        "type": "string",
                        "title": "Contact Person",
                        "default": ""
                    },
                    "contact_tel": {
                        "type": "string",
                        "title": "Contact Tel",
                        "default": ""
                    },
                    "currency": {
                        "type": "string",
                        "title": "Currency",
                        "default": ""
                    },
                    "customer_type": {
                        "type": "string",
                        "title": "Customer Type",
                        "default": ""
                    },
                    "handler": {
                        "type": "string",
                        "title": "Handler",
                        "default": ""
                    },
                    "price_print": {
                        "type": "integer",
                        "title": "Price Print",
                        "default": 1
                    },
                    "payment_amount": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Payment Amount"
                    },
                    "remark": {
                        "type": "string",
                        "title": "Remark",
                        "default": ""
                    },
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/CreateShipmentDetailRow"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "customer_id",
                    "shipment_date",
                    "warehouse",
                    "detail"
                ],
                "title": "CreateShipmentRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ShipmentDetail": {
                "properties": {
                    "main": {
                        "$ref": "#/components/schemas/ShipmentMainInfo"
                    },
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ShipmentDetailRow"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "main",
                    "detail"
                ],
                "title": "ShipmentDetail"
            },
...
            "ShipmentListItem": {
                "properties": {
                    "order_no": {
                        "type": "string",
                        "title": "Order No"
                    },
                    "order_date": {
                        "type": "string",
                        "title": "Order Date"
                    },
                    "customer_id": {
                        "type": "string",
                        "title": "Customer Id"
                    },
                    "customer_addr": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Addr",
                        "default": ""
                    },
                    "state": {
                        "type": "integer",
                        "title": "State"
                    },
                    "total_qty": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Qty",
                        "default": 0
                    },
                    "total_amount": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Amount",
                        "default": 0
                    },
                    "tracking_no": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tracking No",
                        "default": ""
                    },
                    "customer_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Name",
                        "default": ""
                    },
                    "print_count": {
                        "type": "integer",
                        "title": "Print Count",
                        "default": 0
                    },
                    "salesperson": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Salesperson",
                        "default": ""
                    },
                    "shipping_addr": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipping Addr",
                        "default": ""
                    },
                    "shipping_method": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipping Method",
                        "default": ""
                    },
                    "freight": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Freight"
                    }
                },
                "type": "object",
                "required": [
                    "order_no",
                    "order_date",
                    "customer_id",
                    "state"
                ],
                "title": "ShipmentListItem"
            },
            "ShipmentListResponse": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "rows": {
                        "items": {
                            "$ref": "#/components/schemas/ShipmentListItem"
                        },
                        "type": "array",
                        "title": "Rows"
                    }
                },
                "type": "object",
                "required": [
                    "total",
                    "rows"
                ],
                "title": "ShipmentListResponse"
            },
...
            "UpdateShipmentRequest": {
                "properties": {
                    "customer_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Id"
                    },
                    "shipment_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipment Date"
                    },
                    "warehouse": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Warehouse"
                    },
                    "customer_addr": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Addr"
                    },
                    "shipping_addr": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipping Addr"
                    },
                    "shipping_tel": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipping Tel"
                    },
                    "shipping_method": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Shipping Method"
                    },
                    "delivery_person": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Delivery Person"
                    },
                    "tracking_no": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tracking No"
                    },
                    "freight": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Freight"
                    },
                    "salesperson": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Salesperson"
                    },
                    "contact_person": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Contact Person"
                    },
                    "contact_tel": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Contact Tel"
                    },
                    "currency": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Currency"
                    },
                    "customer_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Type"
                    },
                    "handler": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Handler"
                    },
                    "price_print": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Print"
                    },
                    "payment_amount": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Payment Amount"
                    },
                    "remark": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Remark"
                    },
                    "detail": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/CreateShipmentDetailRow"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "UpdateShipmentRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "WriteOperationResponse": {
                "properties": {
                    "dh": {
                        "type": "string",
                        "title": "Dh"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "state": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "State"
                    }
                },
                "type": "object",
                "required": [
                    "dh",
                    "message"
                ],
                "title": "WriteOperationResponse"
            }
        }
    }
}